Refactor devices page components
